% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msm.stacked-package.R
\docType{package}
\name{msm.stacked-package}
\alias{msm.stacked-package}
\title{Stacked Probabilities Plots for 'msm' Multistate Models}
\description{
Create stacked probabilities plots for multistate models fitted using
the 'msm' package. Plots are created using the 'ggplot2' package, and can be
further customised by the user. State occupancy probabilities over time can also
be calculated, including conditional predictions on being in a certain state at
a given time.
}
\author{
Alessandro Gasparini (alessandro.gasparini@reddooranalytics.se)
}
