% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked.data.msm.R
\name{stacked.data.msm}
\alias{stacked.data.msm}
\title{Calculate State Occupancy Probabilities Over Time from and to Every State}
\usage{
stacked.data.msm(model, tstart, tforward, tseqn = 10, ...)
}
\arguments{
\item{model}{A multistate model fitted using \code{\link[msm:msm]{msm::msm()}}.}

\item{tstart}{Numeric value denoting the starting time for calculating predictions.
It can be time zero (e.g., the beginning of the study) or it can be a given follow-up
time (e.g., one year after baseline) in which case predictions will be conditional on
being in a given state at time \code{tstart}.}

\item{tforward}{Numeric value denoting how many units of time forward predictions are
to be calculated for. For instance, if \code{tstart = t0} and \code{tforward = t1}, predictions
will be computed up to time \code{t0 + t1}.}

\item{tseqn}{Numeric value denoting how many sub-intervals to use between \code{tstart} and
\code{tstart + tforward}, defaulting to 10. A larger number of \code{tnseq} will lead to smoother plots once using
the \code{\link[=stacked.plot.msm]{stacked.plot.msm()}} function.}

\item{...}{Additional arguments to be passed to \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}. This is useful,
for example, if calculating predictions for a certain covariates pattern - otherwise,
as in \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}, predictions will be assuming means of the covariates in
the data.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{from}, denoting the starting state;
\item \code{to}, denoting the destination state;
\item \code{p}, denoting the probability of being in state \code{to} starting from state \code{from} at
time \code{t};
\item \code{tstart}, denoting the starting point for the predictions;
\item \code{t}, denoting the times for the predicted probabilities \code{p}.
}
}
\description{
The \code{\link[=stacked.data.msm]{stacked.data.msm()}} function can be used to calculate
state occupancy probabilities over time for multistate models fitted
using the {msm} package.
}
\seealso{
\code{\link[msm:msm]{msm::msm()}}, \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}
}
