% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked.plot.msm.R
\name{stacked.plot.msm}
\alias{stacked.plot.msm}
\title{Produce Stacked Probabilities Plots}
\usage{
stacked.plot.msm(
  ...,
  plab = "From",
  ylab = "Time",
  xlab = "Probability",
  start0 = TRUE
)
}
\arguments{
\item{...}{Arguments passed onto \code{\link[=stacked.data.msm]{stacked.data.msm()}}.}

\item{plab}{A string denoting the label of each facet. Defaults to "From".}

\item{ylab}{A string denoting the label of the vertical axis. Defaults to "Probability".}

\item{xlab}{A string denoting the label of the horizontal axis. Defaults to "Time".}

\item{start0}{A boolean value denoting whether plots should start the horizontal axis
at time zero, irrespectively of \code{tstart}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
The \code{\link[=stacked.plot.msm]{stacked.plot.msm()}} function can be used to produce stacked
probabilities plots for multistate models fitted using the {msm} package.
}
\seealso{
\code{\link[=stacked.data.msm]{stacked.data.msm()}}, \code{\link[msm:msm]{msm::msm()}}, \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}
}
