% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.msm.R
\name{states.msm}
\alias{states.msm}
\title{List States of a Multi-State Model}
\usage{
states.msm(model)
}
\arguments{
\item{model}{An object of class \code{msm}.}
}
\value{
A string vector of state names.
}
\description{
The \code{\link[=states.msm]{states.msm()}} function lists the possible states of a
multi-state model fitted with the {msm} package.
}
\examples{

# Based on example in ?msm
# We first fit a {msm} model without covariates
library(msm)
twoway4.q <- rbind(
  c(-0.5, 0.25, 0, 0.25),
  c(0.166, -0.498, 0.166, 0.166),
  c(0, 0.25, -0.5, 0.25),
  c(0, 0, 0, 0)
)
cav.msm <- msm(
  formula = state ~ years,
  subject = PTNUM,
  data = cav,
  qmatrix = twoway4.q,
  deathexact = 4,
  fixedpars = TRUE # only to speed up examples!
)

# Possible states:
states.msm(cav.msm)
}
