% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked.data.msm.R
\name{stacked.data.msm}
\alias{stacked.data.msm}
\title{Calculate State Occupancy Probabilities Over Time from and to Every State}
\usage{
stacked.data.msm(model, tstart, tforward, tseqn = 5, exclude = NULL, ...)
}
\arguments{
\item{model}{A multi-state model fitted using \code{\link[msm:msm]{msm::msm()}}.}

\item{tstart}{Numeric value denoting the starting time for calculating predictions.
It can be time zero (e.g., the beginning of the study) or it can be a given follow-up
time (e.g., one year after baseline) in which case predictions will be conditional on
being in a given state at time \code{tstart}.}

\item{tforward}{Numeric value denoting how many units of time forward predictions are
to be calculated for. For instance, if \code{tstart = t0} and \code{tforward = t1}, predictions
will be computed up to time \code{t0 + t1}.}

\item{tseqn}{Numeric value denoting how many sub-intervals to use between \code{tstart} and
\code{tstart + tforward}, defaulting to 5. This must be greater than one.}

\item{exclude}{Denotes a state (or more than one) for which transitions from such state are
to be excluded. This is useful, for example, to avoid returning transitions from an absorbing
state; the \code{\link[=states.msm]{states.msm()}} function can be helpful to identify the names of the states for a
given model. The \code{exclude} parameter defaults to \code{NULL}, where all transitions are returned.}

\item{...}{Additional arguments to be passed to \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}. This is useful,
for example, if calculating predictions for a certain covariates pattern - otherwise,
as in \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}, predictions will be assuming means of the covariates in
the data.}
}
\value{
A data frame with the following columns:
\itemize{
\item \code{from}, denoting the starting state;
\item \code{to}, denoting the destination state;
\item \code{p}, denoting the probability of being in state \code{to} starting from state \code{from} at
time \code{t};
\item \code{tstart}, denoting the starting point for the predictions;
\item \code{t}, denoting the times for the predicted probabilities \code{p}.
}
}
\description{
The \code{\link[=stacked.data.msm]{stacked.data.msm()}} function can be used to calculate
state occupancy probabilities over time for multi-state models fitted
using the {msm} package.
}
\examples{

### Example 1:
# Based on example in ?msm
# We first fit a {msm} model without covariates
library(msm)
twoway4.q <- rbind(
  c(-0.5, 0.25, 0, 0.25),
  c(0.166, -0.498, 0.166, 0.166),
  c(0, 0.25, -0.5, 0.25),
  c(0, 0, 0, 0)
)
cav.msm <- msm(
  formula = state ~ years,
  subject = PTNUM,
  data = cav,
  qmatrix = twoway4.q,
  deathexact = 4,
  fixedpars = TRUE # only to speed up examples!
)

# Predictions from time 0 to time 1, with 3 mid-points:
p1 <- stacked.data.msm(model = cav.msm, tstart = 0, tforward = 1, tseqn = 3)
head(p1)

# Predictions from time 1 to time 5, with 5 mid-points:
p2 <- stacked.data.msm(model = cav.msm, tstart = 1, tforward = 4, tseqn = 5)
head(p2)

### Example 2:
# Model with covariates
cav.msm.cov <- msm(
  formula = state ~ years,
  subject = PTNUM,
  data = cav,
  covariates = ~sex,
  qmatrix = twoway4.q,
  deathexact = 4,
  fixedpars = TRUE # only to speed up examples!
)

# Predictions from time 0 to time 5, mean covariates values:
p3 <- stacked.data.msm(model = cav.msm.cov, tstart = 0, tforward = 5)

# Predictions from time 0 to time 5, for `sex = 0`:
p4 <- stacked.data.msm(model = cav.msm.cov, tstart = 0, tforward = 5, covariates = list(sex = 0))

# Predictions from time 0 to time 5, for `sex = 1`:
p5 <- stacked.data.msm(model = cav.msm.cov, tstart = 0, tforward = 5, covariates = list(sex = 1))

# p3, p4, and p5 should all be different:
all.equal(p3, p4)
all.equal(p3, p5)
all.equal(p4, p5)

### Example 3:
# Excluding transitions from a certain state, e.g., from State 4:
stacked.data.msm(model = cav.msm.cov, tstart = 0, tforward = 5, exclude = "State 4")

# Returning transitions from State 1 only:
stacked.data.msm(
  model = cav.msm.cov,
  tstart = 0,
  tforward = 5,
  exclude = c("State 2", "State 3", "State 4")
)
}
\seealso{
\code{\link[msm:msm]{msm::msm()}}, \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}
}
