% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked.plot.msm.R
\name{stacked.plot.msm}
\alias{stacked.plot.msm}
\title{Produce Stacked Probabilities Plots}
\usage{
stacked.plot.msm(
  ...,
  tseqn = 50,
  plab = "From",
  xlab = "Time",
  ylab = "Probability",
  start0 = TRUE
)
}
\arguments{
\item{...}{Arguments passed onto \code{\link[=stacked.data.msm]{stacked.data.msm()}}.}

\item{tseqn}{Numeric value denoting how \emph{smooth} the curves should be over time, defaulting to 50.
Please see documentation for the \code{\link[=stacked.data.msm]{stacked.data.msm()}} function for more details, as it works
in a similar way.}

\item{plab}{A string denoting the label of each facet. Defaults to "From".}

\item{xlab}{A string denoting the label of the horizontal axis. Defaults to "Time".}

\item{ylab}{A string denoting the label of the vertical axis. Defaults to "Probability".}

\item{start0}{A boolean value denoting whether plots should start the horizontal axis
at time zero, irrespectively of \code{tstart}. See \code{\link[=stacked.data.msm]{stacked.data.msm()}} for more details
on the \code{tstart} argument.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
The \code{\link[=stacked.plot.msm]{stacked.plot.msm()}} function can be used to produce stacked
probabilities plots for multi-state model fits from the {msm} package.
}
\examples{

### Example 1:
# Based on example in ?msm
# We first fit a {msm} model without covariates
library(msm)
twoway4.q <- rbind(
  c(-0.5, 0.25, 0, 0.25),
  c(0.166, -0.498, 0.166, 0.166),
  c(0, 0.25, -0.5, 0.25),
  c(0, 0, 0, 0)
)
cav.msm <- msm(
  formula = state ~ years,
  subject = PTNUM,
  data = cav,
  qmatrix = twoway4.q,
  deathexact = 4,
  fixedpars = TRUE # only to speed up examples!
)

# Stacked probablities plot from 0 to 5:
stacked.plot.msm(model = cav.msm, tstart = 0, tforward = 5)
# ...or, with finer predictions:
stacked.plot.msm(model = cav.msm, tstart = 0, tforward = 5, tseqn = 100)

# Predictions starting from time 1 to time 5:
stacked.plot.msm(model = cav.msm, tstart = 1, tforward = (5 - 1))
# Note that by default the plot starts at zero.
# If we wanted to start from `tstart:`
stacked.plot.msm(model = cav.msm, tstart = 1, tforward = (5 - 1), start0 = FALSE)

### Example 2:
# Model with covariates
cav.msm.cov <- msm(
  formula = state ~ years,
  subject = PTNUM,
  data = cav,
  covariates = ~sex,
  qmatrix = twoway4.q,
  deathexact = 4,
  fixedpars = TRUE # only to speed up examples!
)

# Stacked probabilities plot, from time 0 to time 5, mean covariates values:
stacked.plot.msm(model = cav.msm.cov, tstart = 0, tforward = 5)

# Stacked probabilities plot, from time 0 to time 5, for `sex = 0`:
stacked.plot.msm(model = cav.msm.cov, tstart = 0, tforward = 5, covariates = list(sex = 0))

# Stacked probabilities plot, from time 0 to time 5, for `sex = 1`:
stacked.plot.msm(model = cav.msm.cov, tstart = 0, tforward = 5, covariates = list(sex = 1))

# The three plots should be different!

### Example 3:
# We can exclude transitions from a certain state from the plot, e.g., from State 4:
stacked.plot.msm(model = cav.msm, tstart = 0, tforward = 5, exclude = "State 4")
}
\seealso{
\code{\link[=stacked.data.msm]{stacked.data.msm()}}, \code{\link[msm:msm]{msm::msm()}}, \code{\link[msm:pmatrix.msm]{msm::pmatrix.msm()}}
}
